+++ lib/mapplanner/adjacent_test.go
