+++ lib/mapplanner/boundary_wall.go
