+++ lib/mapplanner/connectivity_integration_test.go
