+++ lib/mapplanner/npc_planner.go
