+++ lib/mapplanner/props_planner.go
