+++ lib/mapplanner/rect_room.go
