+++ lib/mapplanner/regeneration_test.go
