+++ lib/mapplanner/room_draw.go
