+++ lib/mapplanner/ruins_planner.go
