+++ lib/mapplanner/single_room.go
