+++ lib/mapplanner/small_room_planner_test.go
