+++ lib/mapplanner/string_map_error_test.go
