+++ lib/mapplanner/string_map_planner_test.go
