+++ lib/mapplanner/string_town_planner_test.go
