+++ lib/mapplanner/validation.go
