+++ lib/mapplanner/wall_type.go
