+++ lib/mapplanner/wall_type_test.go
